renkler

+++ WeAreTheChampions/models/Color.cs
@@ -10,9 +10,9 @@ namespace WeAreTheChampions.models
     {
         public int Id { get; set; }
         public string ColorName { get; set; }
-        public int Red { get; set; }
-        public int Green { get; set; }
-        public int Blue { get; set; }
+        public byte Red { get; set; }
+        public byte Green { get; set; }
+        public byte Blue { get; set; }
         public virtual ICollection<Team> Teams { get; set; }
     }
 }
